* arts-2-0-35
* CMakeLists.txt: Turn on -Wconversion warning by default.

* src/m_cloudbox.cc:
(cloudboxSetAutomatically): Change type of p_margin1 from Index to
Numeric to avoid loss of precision in implicit cast.

(ScatteringParticlesSelect): Change type of intarr_total from Numeric
to Index.

* src/docserver.cc (Docserver::launch): Cast port to uint16_t.

* src/math_funcs.cc, src/m_abs_lookup.cc, src/m_surface.cc:

Add some explicit casts from Index to Numeric to suppress conversion
warning for cases where we know what we're doing.

* src/nc_io_array_types.cc: Remove unused variables.

* src/libmicrohttpd/daemon.c (send_param_adapter): Add explicit cast
to avoid conversion warning.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@7315 aaf1aab0-4228-0410-ad68-8dceda47f409

+++ src/docserver.cc
@@ -27,6 +27,7 @@
 
 #ifdef ENABLE_DOCSERVER
 
+#include <stdint.h>
 #include <string>
 #include <vector>
 #include <sstream>
@@ -2097,7 +2098,7 @@ int Docserver::launch(bool daemon)
   struct MHD_Daemon *d;
   
   d = MHD_start_daemon (MHD_USE_THREAD_PER_CONNECTION | MHD_USE_DEBUG,
-                        mport, NULL, NULL, &ahc_echo, (void *)this, MHD_OPTION_END);
+                        (uint16_t)mport, NULL, NULL, &ahc_echo, (void *)this, MHD_OPTION_END);
   
   if (d == NULL)
   {
